Creating independent functions and freeing space on header file to call only the headers
