+++ README.md
@@ -13,4 +13,4 @@
 ## 🧪 Tests and Error mapping
 
 - To run tests and error mapping. use:
-`gcc -Wall -Wextra -std=gnu99 -fdiagnostics-color=always -Ofast -DVERSION=""2.1.170611.090824"" -DBUILD=""170611.090824"" -DDEBUG=0 -DD_ -D_FORTIFY_SOURCE=1 -Wl,--defsym,BUILD_170611_090824=0 -lm cryptosystem.c -o cryptosystem.x 2>&1 | tee errors.err`
+`gcc -Wall -Wextra -std=gnu18 -fdiagnostics-color=always -Ofast -DVERSION=""2.1.170611.090824"" -DBUILD=""170611.090824"" -DDEBUG=0 -DD_ -D_FORTIFY_SOURCE=1 -Wl,--defsym,BUILD_170611_090824=0 -lm cryptosystem.c -o cryptosystem.x 2>&1 | tee errors.err`
