+++ functions/caesarDecrypt.c
@@ -1,0 +1,39 @@
+int caesarDecrypt()
+{
+    char message[255], ch;
+	int i, key;
+	
+	printf("Input the message to be decrypted: ");
+	fgets(message, 255, stdin);
+	printf("Input the key for decryption: ");
+	scanf("%d", &key);
+	
+	for(i = 0; message[i] != '\0'; ++i){
+		ch = message[i];
+		
+		if(ch >= 'a' && ch <= 'z'){
+			ch = ch - key;
+			
+			if(ch < 'a'){
+				ch = ch + 'z' - 'a' + 1;
+			}
+			
+			message[i] = ch;
+		}
+		else if(ch >= 'A' && ch <= 'Z'){
+			ch = ch - key;
+			
+			if(ch < 'A'){
+				ch = ch + 'Z' - 'A' + 1;
+			}
+			
+			message[i] = ch;
+		}
+	}
+	
+	printf("Decrypted message: %s", message);
+	getchar();
+	printf("\n");
+	
+	return 0;
+}
